chore: add filters for getByItems

+++ dipdup-core/src/main/kotlin/com/rarible/dipdup/client/core/model/DipDupOrder.kt
@@ -7,7 +7,7 @@ import java.time.OffsetDateTime
 data class DipDupOrder(
     val id: String,
     val fill: BigDecimal,
-    val makeStock: BigInteger,
+    val makeStock: BigDecimal,
     val platform: TezosPlatform,
     val cancelled: Boolean,
     val status: OrderStatus,
